Create Composer Upstream Lock Plugin
This creates a composer plugin in `tools/composer-upstream-lock` which
allows specifying an external composer.lock file to lock known packages
to a specific version. This allows rebuilding a composer.json file using
automated systems such as Musa from an empty template without having to
use the exact copies of the composer.* files included in Cable Car.

When a composer.lock file is specified using
`COMPOSER_UPSTREAM_LOCK_FILE` the plugin will limit the installable
version of packages in `composer require` and `composer install` (for
fresh installs) to the specific versions in the specified lock file.
This allos Musa to create sub-sets of the Open Social product.